moved pred model fn to modeling file
